requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Werkzeug==2.3.7
stripe==7.2.0
requests==2.31.0
python-dotenv==1.0.0
discord.py==2.3.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
